requirements.txt
# Core
streamlit==1.39.0
python-docx==1.1.2
Pillow==10.4.0

# Google Drive
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# Dropbox
dropbox==12.0.2
